services/ai-inference/requirements.txt
# AI Inference Service Dependencies
# Python 3.9+

# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Kafka
kafka-python==2.0.3

# MinIO (S3-compatible)
minio==7.2.0

# YOLOv8
ultralytics==8.0.196

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Logging
python-json-logger==2.0.7

# HTTP Client
requests==2.31.0

# System Info
psutil==5.9.6



